将核心代码和UI界面分离，方便后续跨平台支持
