+++ FrpGUI.sln
@@ -1,14 +1,16 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34316.72
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FrpGUI", "FrpGUI\FrpGUI.csproj", "{F6F6CA34-96A5-4E6B-8AD8-E1452288C59A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FrpGUI.WPF", "FrpGUI.WPF\FrpGUI.WPF.csproj", "{F6F6CA34-96A5-4E6B-8AD8-E1452288C59A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{38A08A1F-58A3-47E8-ACDC-CFB555D0FC72}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FrpGUI", "FrpGUI\FrpGUI.csproj", "{DD6722FB-FE07-44AD-92BD-60EC33553F02}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,6 +22,10 @@ Global
 		{F6F6CA34-96A5-4E6B-8AD8-E1452288C59A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F6F6CA34-96A5-4E6B-8AD8-E1452288C59A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F6F6CA34-96A5-4E6B-8AD8-E1452288C59A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DD6722FB-FE07-44AD-92BD-60EC33553F02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DD6722FB-FE07-44AD-92BD-60EC33553F02}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DD6722FB-FE07-44AD-92BD-60EC33553F02}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DD6722FB-FE07-44AD-92BD-60EC33553F02}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
