+++ FrpGUI/Config/IToFrpConfig.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel;
+
+namespace FrpGUI.Config
+{
+    public interface IToFrpConfig : INotifyPropertyChanged
+    {
+        public string ToIni();
+    }
+}
