--- FrpGUI/Config/IToIni.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace FrpGUI
-{
-    public interface IToIni : INotifyPropertyChanged
-    {
-        public string ToIni();
-    }
-}
