--- FrpGUI/Panel/AddRulePanel.xaml
@@ -1,82 +1,0 @@
-﻿<UserControl
-    x:Class="FrpGUI.AddRulePanel"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:FrpGUI"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:me="clr-namespace:FzLib.WPF.MarkupExtensions;assembly=FzCoreLib.Windows"
-    xmlns:ui="http://schemas.modernwpf.com/2019"
-    d:DesignHeight="450"
-    d:DesignWidth="800"
-    mc:Ignorable="d">
-    <UserControl.Resources>
-        <local:RuleParameterEnableConverter x:Key="enableConverter" />
-    </UserControl.Resources>
-    <DockPanel LastChildFill="True">
-        <ui:SimpleStackPanel
-            HorizontalAlignment="Right"
-            DockPanel.Dock="Bottom"
-            Orientation="Horizontal"
-            Spacing="8">
-            <Button
-                Click="SaveButton_Click"
-                Content="保存" />
-            <Button
-                Click="CancelButton_Click"
-                Content="取消" />
-        </ui:SimpleStackPanel>
-        <WrapPanel Margin="8">
-            <WrapPanel.Resources>
-                <Style
-                    BasedOn="{StaticResource DefaultTextBoxStyle}"
-                    TargetType="{x:Type TextBox}">
-                    <Setter Property="Margin" Value="8,8" />
-                    <Setter Property="Width" Value="180" />
-                </Style>
-                <Style
-                    BasedOn="{StaticResource DefaultComboBoxStyle}"
-                    TargetType="{x:Type ComboBox}">
-                    <Setter Property="Margin" Value="8,8" />
-                    <Setter Property="Width" Value="180" />
-                </Style>
-            </WrapPanel.Resources>
-            <TextBox
-                ui:ControlHelper.Header="名称"
-                Text="{Binding Rule.Name}" />
-            <ComboBox
-                ui:ControlHelper.Header="类型"
-                ItemsSource="{me:EnumValues EnumType=local:NetType}"
-                SelectedItem="{Binding Rule.Type}" />
-
-            <TextBox
-                ui:ControlHelper.Header="本地地址"
-                Text="{Binding Rule.LocalAddress}" />
-            <TextBox
-                ui:ControlHelper.Header="本地端口"
-                Text="{Binding Rule.LocalPort}" />
-            <TextBox
-                ui:ControlHelper.Header="远程端口"
-                IsEnabled="{Binding Rule.Type, ConverterParameter=RemotePort, Converter={StaticResource enableConverter}}"
-                Text="{Binding Rule.RemotePort}" />
-            <TextBox
-                ui:ControlHelper.Header="域名"
-                IsEnabled="{Binding Rule.Type, ConverterParameter=Domains, Converter={StaticResource enableConverter}}"
-                Text="{Binding Rule.Domains}" />
-            <TextBox
-                ui:ControlHelper.Header="STCP密钥"
-                IsEnabled="{Binding Rule.Type, ConverterParameter=STCPKey, Converter={StaticResource enableConverter}}"
-                Text="{Binding Rule.STCPKey}" />
-            <TextBox
-                ui:ControlHelper.Header="STCP服务名"
-                IsEnabled="{Binding Rule.Type, ConverterParameter=STCPServerName, Converter={StaticResource enableConverter}}"
-                Text="{Binding Rule.STCPServerName}" />
-            <CheckBox
-                Content="压缩"
-                IsChecked="{Binding Rule.Compression}" />
-            <CheckBox
-                Content="加密"
-                IsChecked="{Binding Rule.Encryption}" />
-        </WrapPanel>
-    </DockPanel>
-</UserControl>
