--- FrpGUI/Panel/ProcessStatus.cs
@@ -1,9 +1,0 @@
-﻿namespace FrpGUI
-{
-    public enum ProcessStatus
-    {
-        NotRun,
-        Running,
-        Busy
-    }
-}
