+++ FrpGUI/ProcessStatus.cs
@@ -1,0 +1,9 @@
+﻿namespace FrpGUI
+{
+    public enum ProcessStatus
+    {
+        NotRun,
+        Running,
+        Busy
+    }
+}
