--- FrpGUI/Properties/launchSettings.json
@@ -1,7 +1,0 @@
-{
-  "profiles": {
-    "FrpGUI": {
-      "commandName": "Project"
-    }
-  }
-}
