+++ FrpGUI/Util/HttpHelper.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using FrpGUI.Config;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
