+++ FrpGUI/Util/MessageEventArgs.cs
@@ -1,0 +1,9 @@
+﻿//namespace FrpGUI.Util
+//{
+//    public class MessageEventArgs(string message) : EventArgs
+//    {
+//        public string Message { get; init; } = message;
+
+//        public static implicit operator MessageEventArgs(string message) { return new MessageEventArgs(message); }
+//    }
+//}
