--- FrpGUI/frp/frps.ini
@@ -1,2 +1,0 @@
-[common]
-bind_port = 7000
